switched priority to int64_t, added task test & fixed task_queue bugs.
Note the task test has most of the coverage but doesn't do boundary/corner case testing yet. Shall be revisited after we have useful scheduler tests.

+++ include/monad/fiber/current.h
@@ -7,7 +7,7 @@ typedef struct monad_fiber
 {
     monad_fiber_task_t task;
     monad_fiber_context_t *context;
-    size_t priority;
+    int64_t priority;
     monad_fiber_scheduler_t *scheduler;
 } monad_fiber_t;
