+++ include/monad/fiber/scheduler.h
@@ -41,12 +41,12 @@ void monad_fiber_scheduler_destroy(monad_fiber_scheduler_t *);
 void monad_fiber_scheduler_stop(monad_fiber_scheduler_t *);
 
 void monad_fiber_scheduler_post(
-    monad_fiber_scheduler_t *, monad_fiber_task_t *task, size_t priority);
+    monad_fiber_scheduler_t *, monad_fiber_task_t *task, int64_t priority);
 void monad_fiber_scheduler_dispatch(
-    monad_fiber_scheduler_t *, monad_fiber_task_t *task, size_t priority);
+    monad_fiber_scheduler_t *, monad_fiber_task_t *task, int64_t priority);
 
 monad_fiber_task_t *monad_fiber_scheduler_pop_higher_priority_task(
-    monad_fiber_scheduler_t *, size_t priority);
+    monad_fiber_scheduler_t *, int64_t priority);
 
 bool monad_fiber_run_one(monad_fiber_scheduler_t *);
